work

+++ FirmwareNew/sd/integer.h
@@ -13,8 +13,8 @@
 #else			/* Embedded platform */
 
 /* These types must be 16-bit, 32-bit or larger integer */
-typedef int				INT;
-typedef unsigned int	UINT;
+typedef signed long     INT;
+typedef unsigned long   UINT;
 
 /* These types must be 8-bit integer */
 typedef char			CHAR;
